Viewer2D: Update list of loaded node's outputs when they are en/dis-abled
When a node with viewable output attributes is loaded, the combo box  of
the 2D Viewer lists these attributes. As some of these attributes are
enabled or disabled, they need to be added or removed from that list,
without needing to load another node first, or to unload the node.

This commit connects the signal that is emitted when an output attribute
with a viewable output is enabled or disabled to the slot that reloads
the currently loaded node (and thus updates the list).